Changed the linked_lists to handle void pointers
Makefile.am :
Added the files to the compilation units and added
the new unit test file to the test runner.

linked_list.c :
Moved the documentation of the functions to the header
file and updated the function to use void pointers
and previous pointers in the list

linked_list.h :
Moved the documentation here, addded a previous pointer
to the linked list and modified the data int to a void
pointer

memory_utils.c:

Made the safe malloc function static.

memory_utils.h :

Made the safe malloc function static and added correct imports
for the the compiler not to warn so much :(

linked_list_test.c :

Added basic unit test to the linked lists, this need to
be drasticly improved as we build on these datastructures.